backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
aiofiles==23.2.1
python-docx==1.1.0
pypdf2==3.0.1
markdown==3.5.2
python-dateutil==2.8.2
requests==2.31.0
PyJWT==2.8.0

